rename a bunch of stuff for typescript
