+++ apps/web/vite.config.ts
@@ -12,7 +12,7 @@ export default defineConfig({
     alias: {
       '@cuptrail/core': path.resolve(__dirname, '../../packages/core'),
       '@cuptrail/utils': path.resolve(__dirname, '../../packages/utils'),
-      '@cuptrail/maps-api': path.resolve(__dirname, '../../packages/maps-api'),
+      '@cuptrail/maps-api': path.resolve(__dirname, '../../packages/maps'),
       '@utils': path.resolve(__dirname, './app/utils'),
       '@components': path.resolve(__dirname, './app/components'),
     },
