+++ packages/maps/normalize.ts
